backend/requirements.txt
# Core dependencies
fastapi==0.115.6
uvicorn==0.24.0
sqlalchemy>=1.4.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.10.4
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
redis==5.0.1
httpx==0.25.2
filetype==1.2.0
aiofiles==23.2.1

# RAG system dependencies
openai>=1.0.0
tiktoken>=0.5.0
numpy>=1.21.0
